+++ elmyra.ip.access.epo/elmyra/ip/util/numbers/numberlists.py
@@ -11,11 +11,13 @@ def parse_numberlist(rawdata):
 
 def normalize_numbers(entries):
     entries = map(lambda s: s.replace(u' ', u''), entries)
-    response = {'valid': [], 'invalid': []}
+    response = {'valid': [], 'invalid': [], 'all': []}
     for entry in entries:
         entry_normalized = normalize_patent(entry, fix_kindcode=True)
         if entry_normalized:
             response['valid'].append(entry_normalized)
+            response['all'].append(entry_normalized)
         else:
             response['invalid'].append(entry)
+            response['all'].append(entry)
     return response